GCP/cloud-function/atualiza-dados-nubank/requirements.txt
# Function dependencies, for example:
# package>=version

pandas>=1.3.5
pynubank>=2.17.0
numpy>=1.22.3
google-cloud-storage>=2.1.0
google-cloud-bigquery>=2.34.2
google-cloud-secret-manager==2.10.0